Update RadarDataser.cs Data varariable as int
Data variable in RadarDataset class of the RadarChartComponent is a double type list. Instead of doing a _rand.NextDouble() to assign the chart distribution double values, I would like to give int values for distribution on radar chart. But the Data variable defined in the RadarDataset class only takes a list with double type.